Rare Candy Can Evolve Level 100 Pokémon
Not the best implementation though. I tried Rangi's but it would level the Pokémon to level 1, then evolve, then return back to level 100. My version doesn't level to level 1 but it doesn't check if the Pokémon is already at it's max evolved form because I don't know how to do that 😬

https://github.com/Rangi42/polishedcrystal/commit/384a38aa71285146338ac4b8af281b71b2f65724